backend/requirements.txt
django==1.11.3

djangorestframework==3.6.3

django-filter==1.0.4
django-cors-headers==2.0.1